move common logic of AnnotationManagement to common directory
